Fix printf format for orig_rax in trace.c
Change printf format specifier for orig_rax to match its type.

+++ strace/trace.c
@@ -26,7 +26,7 @@ void print_syscall(pid_t child)
 	struct user_regs_struct regs;
 
 	if (ptrace(PTRACE_GETREGS, child, NULL, &regs) != -1)
-		printf("%llu\n", regs.orig_rax);
+		printf("%lu\n", (unsigned long)regs.orig_rax);
 }
 
 /**
